Changed delay from time steps to ms (as long as delta t is 1ms, no change in current code is required). Bug fix in STDPConn to grab activity from pre layer with respect to delay. Added init delay to InitUniformRandomWeights
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@5666 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/HyPerConn.hpp
@@ -108,7 +108,7 @@ public:
       return weightInitializer;
    }
 
-   void setDelay(int axonId, int delay);
+   void setDelay(int axonId, float delay);
 
    inline int getDelay(int arborId = 0) {
       assert(arborId >= 0 && arborId < numAxonalArborLists);
